refactored code placement
