--- src/javascripts/helpers.js/buttonEvents.js
@@ -1,8 +1,0 @@
-import { feedHealthyFood, feedUnHealthyFood } from '../eat';
-
-const buttonEvents = () => {
-  document.querySelector('#eat-healthy-btn').addEventListener('click', feedHealthyFood);
-  document.querySelector('#eat-unhealthy-btn').addEventListener('click', feedUnHealthyFood);
-};
-
-export default buttonEvents;
