+++ src/javascripts/helpers/renderToDom.js
