+++ src/javascripts/main.js
@@ -1,5 +1,5 @@
 import '../styles/main.scss';
-import { eatCard } from './eat';
+import eatCard from './eat';
 import initialScreen from './initialScreen';
 
 const init = () => {
